--- src/utils.js
@@ -1,7 +1,0 @@
-export const removeHiddenClass = ($selector) => {
-  $selector.classList.remove("hidden");
-};
-
-export const displayTemplate = ($selector, template) => {
-  $selector.innerHTML = template;
-};
